refactor: move plant plotting to dedicated file
